backend/requirements.txt
# FastAPI and Server
fastapi
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.18
asyncpg>=0.29.0
requests>=2.31.0
alembic>=1.13.1

# Redis and Caching
redis>=5.0.1
hiredis>=2.3.2

# AI/ML Libraries
groq>=0.4.2
huggingface-hub>=0.20.0
langchain
langchain-community
langchain-openai>=0.0.5
sentence-transformers>=2.3.1
faiss-cpu>=1.9.0
torch>=2.0.0
chromadb
neo4j>=5.14.0
wikipedia-api>=0.6.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
spacy>=3.7.2

# Local Model Support
transformers
accelerate>=0.25.0
bitsandbytes>=0.41.0
safetensors>=0.4.0

# OpenAI API
openai>=1.10.0
google-generativeai>=0.4.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Utilities
httpx>=0.26.0
tenacity>=8.2.3
python-json-logger>=2.0.7
psutil>=5.9.0

# Monitoring and Observability
prometheus-client>=0.19.0

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
black>=24.1.1
flake8>=7.0.0
mypy>=1.8.0

# RL Optimizer
numpy>=1.26.0
matplotlib>=3.8.0
tqdm>=4.66.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
